+++ src/codeFixers/AlCodeActionProvider.js
@@ -4,6 +4,9 @@ const AA0008 = require('./AA0008');
 const AA0139 = require('./AA0139');
 const AL0666 = require('./AL0666');
 const PRAGMA = require('./pragma');
+const PRAGMAWITHTODO = require('./pragmaWithTodo');
+const PRAGMAALL = require('./pragmaAll');
+const constants = require('../constants');
 
 exports.AlCodeActionProvider = class AlCodeActionProvider {
     relevantDiagnostics;
@@ -18,7 +21,9 @@ exports.AlCodeActionProvider = class AlCodeActionProvider {
             'AA0008': new AA0008.MissingBracketsCodeFixer(context, 'AA0008'),
             'AA0139': new AA0139.PossibleOverflowCodeFixer(context, 'AA0139'),
             'AL0666': new AL0666.RegionFixer(context, 'AL0666'),
-            'PRAGMA': new PRAGMA.AddPragmaCodeFixer(context, 'PRAGMA')
+            'PRAGMA': new PRAGMA.AddPragmaCodeFixer(context, 'PRAGMA'),
+            'PRAGMAALL': new PRAGMAALL.AddPragmaAllCodeFixer(context, constants.PragmaAll),
+            'PRAGMAWITHTODO': new PRAGMAWITHTODO.AddPragmaWithTodoCodeFixer(context, 'PRAGMAWITHTODO')
         }
         this.relevantDiagnostics = Object.keys(this.diagnosticCodeToFix);
     }
@@ -42,9 +47,13 @@ exports.AlCodeActionProvider = class AlCodeActionProvider {
         let warnings = context.diagnostics.filter(diagnostic=>[diagnostic.severity==vscode.DiagnosticSeverity.Warning,diagnostic.severity==vscode.DiagnosticSeverity.Information]);
         warnings = warnings.filter(warning=> !["AA0021","AL0604","AA0139"].includes(warning.code.toString()));
         const fix = this.diagnosticCodeToFix["PRAGMA"];
+        const fixAll = this.diagnosticCodeToFix["PRAGMAALL"];
+        const fixWithTodo = this.diagnosticCodeToFix["PRAGMAWITHTODO"];
         for(let i = 0; i < warnings.length; i++) {
             if (token.isCancellationRequested) return;
-            actions = actions.concat([makeCodeAction(fix.title.replace("{0}",warnings[i].code), fix.commandName, codeFixer.fixTypes.Once, warnings[i], document)]);
+            actions = actions.concat([makeCodeAction(fix.title.replace("{0}",warnings[i].code), fix.commandName, codeFixer.fixTypes.Once, warnings[i], document)]);            
+            actions = actions.concat([makeCodeAction(fixWithTodo.title.replace("{0}",warnings[i].code), fixWithTodo.commandName, codeFixer.fixTypes.Once, warnings[i], document)]);
+            actions = actions.concat([makeCodeAction(fixAll.title.replace("{0}",warnings[i].code) + " (in all documents)", fixAll.commandName, codeFixer.fixTypes.AllDocuments, warnings[i], document)]);            
         }
         return actions;
     }