+++ src/fileManagement/alFileManagement.js
@@ -154,7 +154,6 @@ exports.getFileFormatForType = function getFileFormatForType(alObjectType) {
     return fileLocationFormat;
 };
 
-const alObjectRegex = /^\s*(?<type>\w+)(\s+(?<id>\d+))?\s+(?<name>\w+|"[^"]*")(\s+extends\s+(?<extendedName>\w+|"[^"]*")(\s*\/\/\s*(?<extendedId>\d+))?|\s+implements\s+(?<interfaces>(\w+|"[^"]*")(\s*,\s*(\w+|"[^"]*"))*))?/im;
 class AlObjectInfo {
     type;
     id;
@@ -170,7 +169,7 @@ class AlObjectInfo {
      */
     constructor(textDocument) {
         this.path = textDocument.uri;
-        const match = alObjectRegex.exec(textDocument.getText());
+        const match = constants.AlObjectRegex.exec(textDocument.getText());
         if (match === null) return;
         this.type = match.groups.type.toLowerCase();
         if(Object.values(constants.AlObjectTypes).includes(this.type)){