+++ CHANGELOG.md
@@ -114,3 +114,13 @@ Added Command:
 
 ## [2.0.14]
 - [Prefix Objects adds too many double quotes](https://github.com/BartPermentier/al-toolbox/issues/44)
+
+## [2.0.15]
+- Optimized SetLoadFields functionality
+  - Improved record parameter support  #49
+  - Exclude SystemId, RecordId and several other system functions #51
+  - Only add unique fields (case-insensitive) #50
+  - Surround single if ... then or else with begin..end #45
+  - Fixed issue on local records of the same type as the current object #48
+- Added CodeAction to Add Pragma with TODO Tag #46 & #53
+- Added CodeAction to Add Pragma for all error occurences in all documents #42